src/requirements.txt
flask>=1.1
waitress
requests
pyyaml
crontab
pytz